+++ packages/omgx/wallet/deployment/local/addresses.json
@@ -1,12 +1,12 @@
 {
-  "L1LiquidityPool": "0x6353d6c697cc36446EC727Ae032B518b7FbEc82d",
-  "L2LiquidityPool": "0x3e4CFaa8730092552d9425575E49bB542e329981",
-  "L1ERC20": "0x6C024eeE101AdB6b5d211947fD4B7f6182042978",
-  "L2DepositedERC20": "0x65F72DF8a668BC6272B059BB7F53ADc91066540C",
-  "L1ERC20Gateway": "0x3C38fd8D4e647b230D006d98793Fa3A0ec60bC41",
+  "L1LiquidityPool": "0x195d3769BDA68410F5FB3aa79f2DC4D7Cf67befe",
+  "L2LiquidityPool": "0xfB995f8fd71A30b7A0F4C5Ad98aB1FCe0a7E9Ca2",
+  "L1ERC20": "0xDB1b9ffB4c6630cAEB032e1F4dAe928B167CeceF",
+  "L2DepositedERC20": "0xE607Cca41032d7Ca667b7F45e6138913a087E600",
+  "L1ERC20Gateway": "0xDbA7Ebf2972cFe4B6632fc42f86118556fC41704",
   "l1ETHGatewayAddress": "0x4F53A01556Dc6f120db9a1b4caE786D5f0b5792C",
   "l1MessengerAddress": "0xA6404B184Ad3f6F41b6472f02ba45f25C6A66d49",
-  "L2TokenPool": "0xcD46d1D296466e7F2842A718576d49576fe76a9f",
-  "AtomicSwap": "0xB3FdF320A0bb4b9eBedf4726F0C9Bf49D1268cf9",
-  "L2ERC721": "0x5bD12f0B56C8973ac98446435E07Df8311Aa362c"
+  "L2TokenPool": "0xCe92db3E3B7fdB65f7649c852Cd1E68283d2f4e5",
+  "AtomicSwap": "0x73D9B276944E59A9dDd9369e1c9058e4c57A9B09",
+  "L2ERC721": "0xeEb6b2a4EFA267e100699505FBB4254492773300"
 }
